+++ modules/dynamodb.tf
@@ -8,8 +8,8 @@ locals {
 
 resource "aws_dynamodb_table" "accounts" {
   name             = "Accounts${local.table_suffix}"
-  read_capacity    = 5
-  write_capacity   = 5
+  read_capacity    = var.accounts_table_rcu
+  write_capacity   = var.accounts_table_wcu
   hash_key         = "Id"
   stream_enabled   = true
   stream_view_type = "NEW_AND_OLD_IMAGES"
@@ -18,8 +18,8 @@ resource "aws_dynamodb_table" "accounts" {
     name            = "AccountStatus"
     hash_key        = "AccountStatus"
     projection_type = "ALL"
-    read_capacity   = 5
-    write_capacity  = 5
+    read_capacity   = var.accounts_table_rcu
+    write_capacity  = var.accounts_table_wcu
   }
 
   server_side_encryption {
@@ -52,8 +52,8 @@ resource "aws_dynamodb_table" "accounts" {
 
 resource "aws_dynamodb_table" "leases" {
   name             = "Leases${local.table_suffix}"
-  read_capacity    = 5
-  write_capacity   = 5
+  read_capacity    = var.leases_table_rcu
+  write_capacity   = var.leases_table_wcu
   hash_key         = "AccountId"
   range_key        = "PrincipalId"
   stream_enabled   = true
@@ -67,24 +67,24 @@ resource "aws_dynamodb_table" "leases" {
     name            = "PrincipalId"
     hash_key        = "PrincipalId"
     projection_type = "ALL"
-    read_capacity   = 5
-    write_capacity  = 5
+    read_capacity   = var.leases_table_rcu
+    write_capacity  = var.leases_table_wcu
   }
 
   global_secondary_index {
     name            = "LeaseStatus"
     hash_key        = "LeaseStatus"
     projection_type = "ALL"
-    read_capacity   = 5
-    write_capacity  = 5
+    read_capacity   = var.leases_table_rcu
+    write_capacity  = var.leases_table_wcu
   }
 
   global_secondary_index {
     name            = "LeaseId"
     hash_key        = "Id"
     projection_type = "ALL"
-    read_capacity   = 5
-    write_capacity  = 5
+    read_capacity   = var.leases_table_rcu
+    write_capacity  = var.leases_table_wcu
   }
 
   # AWS Account ID
@@ -126,8 +126,8 @@ resource "aws_dynamodb_table" "leases" {
 
 resource "aws_dynamodb_table" "usage" {
   name             = "Usage${local.table_suffix}"
-  read_capacity    = 5
-  write_capacity   = 5
+  read_capacity    = var.usage_table_rcu
+  write_capacity   = var.usage_table_wcu
   hash_key         = "StartDate"
   range_key        = "PrincipalId"
   stream_enabled   = true